refactor: split LearningHeaderActionsSlot into v1/v2 subfolders
