+++ src/plugin-slots/LearningHeaderActionsSlot/v1/index.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { PluginSlot } from '@openedx/frontend-plugin-framework';
+import HeaderNotificationsSlot from '../../HeaderNotificationsSlot';
+import LearningHelpSlot from '../../LearningHelpSlot';
+
+const LearningHeaderActionsSlotV1 = () => (
+  <PluginSlot
+    id="org.openedx.frontend.layout.learning_header_actions.v1"
+  >
+    <HeaderNotificationsSlot />
+    <LearningHelpSlot />
+  </PluginSlot>
+);
+
+export default LearningHeaderActionsSlotV1;
